requirements.txt
Flask
python-dotenv
replicate
cloudinary